test for handling Ticket Process works well
checking if the modal openes on click. then inputing the information. submiting, checking to see the list counter has changed, switching to the handled list and checking that the information inside the fields is identical to the information i filled out. , then opening it up again and sending the information back(reopening)